requirements.txt
Django>=5.1.0
djangorestframework>=3.15.0
django-environ>=0.11.0
requests>=2.32.0
google-generativeai>=0.8.0
ollama>=0.6.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
python-json-logger>=2.0.7
pytest-django>=4.9.0
pytest-mock>=3.14.0
